requirements.txt
boto3
pandas
snowflake-connector-python
